+++ es7/index.js
@@ -1,4 +1,6 @@
 require('../modules/es7.array.includes');
+require('../modules/es7.array.flat-map');
+require('../modules/es7.array.flatten');
 require('../modules/es7.string.at');
 require('../modules/es7.string.pad-start');
 require('../modules/es7.string.pad-end');