+++ fn/array/index.js
@@ -21,4 +21,6 @@ require('../../modules/es6.array.find-index');
 require('../../modules/es6.array.species');
 require('../../modules/es6.array.iterator');
 require('../../modules/es7.array.includes');
+require('../../modules/es7.array.flat-map');
+require('../../modules/es7.array.flatten');
 module.exports = require('../../modules/_core').Array;